requirements.txt
# Install requirements
Flask==2.1.2
gantry==0.4.15
jupyterlab==3.4.3
loguru==0.6.0
numpy>=1.19.0
pandas>=1.1.0
python-dotenv==0.20.0
scikit-learn==0.24.1
streamlit==1.10.0
